refactor(language-server): Extract handlers to separate files
Breaks up the large session.ts file into separate handlers and consolidates the code related to completions in the server into the handler file
